fix solutions compilation error

+++ include/dynoplan/tdbastar/tdbastar.hpp
@@ -52,9 +52,9 @@ struct AStarNode {
   const ob::State *state;
   Eigen::VectorXd state_eig;
 
-  float fScore;
-  float gScore;
-  float hScore;
+  double fScore;
+  double gScore;
+  double hScore;
   int focalHeuristic; // ecbs
   std::vector<int> motions{}; // list of applicable motions
 
@@ -73,7 +73,7 @@ struct AStarNode {
   // can arrive at this node at time gScore, starting from came_from, using
   // motion used_motion
   struct arrival {
-    float gScore;
+    double gScore;
     // AStarNode* came_from;
     std::shared_ptr<AStarNode> came_from;
     size_t used_motion;
